WIP moving Automatic Manipulator Control stuff into own asmdef
