+++ Assets/Scripts/AutomaticManipulatorControl.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 270002354b7987048bdd125807665e4d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
