+++ Assets/Scripts/AutomaticManipulatorControl/AutomaticManipulatorControlHandler.cs.meta
