+++ Assets/Scripts/AutomaticManipulatorControl/trajectoryplanner.automaticmanipulatorcontrol.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: b23542acc15ade043a49cf34d257520c
-folderAsset: yes
-DefaultImporter:
+guid: 8fbc27192599cbd42a801ae7fb0779fb
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
