+++ Assets/Scripts/TrajectoryPlanner/trajectoryplanner.core.asmdef
@@ -18,7 +18,8 @@
         "GUID:a159cd9c2fa32b146a6ef7abc9fd3cfd",
         "GUID:bd47309b6ac083b41a65d22a204fec8c",
         "GUID:a99f58d8f2dd9754d8f246b5299f912e",
-        "GUID:d7a587e28a05765459779647806c4618"
+        "GUID:d7a587e28a05765459779647806c4618",
+        "GUID:8fbc27192599cbd42a801ae7fb0779fb"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
